Refactored trip data code for error message and added countdown logic- Refactored the data error validation code and moved it to the  updateSaveTripUI function. Introduced countdown logic code to the create save trip function.

+++ src/client/js/lib/trip/removeSavedTrip.js
@@ -1,9 +1,15 @@
 const $WEATHER_SAVE_ELEMENT = document.querySelector('.saved-trips');
 
 const removeSavedTrip = async(e)=>{
-    
-    const data = await Client.deleteSavedTrip(e.target.id);
-    Client.updateSaveTripUI(data);
+
+    if(e.target.classList.contains('save-trip-remove'))
+    {
+        const data = await Client.deleteSavedTrip(e.target.id);
+        Client.updateSaveTripUI(data);
+
+
+
+    }
    
 }
